requirements.txt
langchain
python_dotenv
cohere
pypdf
langchain-cohere
faiss-cpu
langchain-community
langchain-cohere
streamlit

